+++ backend_test.go
@@ -19,7 +19,7 @@ type mockEventsSender struct {
 	eventsProcessed []*logical.EventReceived
 }
 
-func (m *mockEventsSender) Send(ctx context.Context, eventType logical.EventType, event *logical.EventData) error {
+func (m *mockEventsSender) SendEvent(ctx context.Context, eventType logical.EventType, event *logical.EventData) error {
 	if m == nil {
 		return nil
 	}
@@ -41,7 +41,7 @@ func (m *mockEventsSender) expectEvents(t *testing.T, expectedEvents []expectedE
 		if expected.eventType != actual.EventType {
 			t.Fatalf("Mismatched event type at index %d. Expected %s, got %s\n%v", i, expected.eventType, actual.EventType, m.eventsProcessed)
 		}
-		actualPath := actual.Event.Metadata.Fields["path"].GetStringValue()
+		actualPath := actual.Event.Metadata.Fields["secret_path"].GetStringValue()
 		if expected.path != actualPath {
 			t.Fatalf("Mismatched path at index %d. Expected %s, got %s\n%v", i, expected.path, actualPath, m.eventsProcessed)
 		}