+++ path_config.go
@@ -174,8 +174,7 @@ func (b *versionedKVBackend) pathConfigWrite() framework.OperationFunc {
 		defer b.globalConfigLock.Unlock()
 
 		b.globalConfig = config
-		kvEvent(ctx, b.Backend, 2, "config-write",
-			"path", "config")
+		kvEvent(ctx, b.Backend, "config-write", configPath, configPath, 2)
 		return nil, nil
 	}
 }