+++ path_metadata.go
@@ -416,9 +416,7 @@ func (b *versionedKVBackend) pathMetadataWrite() framework.OperationFunc {
 		}
 
 		err = b.writeKeyMetadata(ctx, req.Storage, meta)
-		kvEvent(ctx, b.Backend, 2, "metadata-write",
-			"path", "metadata/"+key,
-		)
+		kvEvent(ctx, b.Backend, "metadata-write", key, "metadata/"+key, 2)
 		return resp, err
 	}
 }
@@ -531,9 +529,7 @@ func (b *versionedKVBackend) pathMetadataPatch() framework.OperationFunc {
 			return nil, err
 		}
 
-		kvEvent(ctx, b.Backend, 2, "metadata-patch",
-			"path", "metadata/"+key,
-		)
+		kvEvent(ctx, b.Backend, "metadata-patch", key, "metadata/"+key, 2)
 		return resp, nil
 	}
 }
@@ -577,9 +573,7 @@ func (b *versionedKVBackend) pathMetadataDelete() framework.OperationFunc {
 
 		// Use encrypted key storage to delete the key
 		err = es.Delete(ctx, key)
-		kvEvent(ctx, b.Backend, 2, "metadata-delete",
-			"path", "metadata/"+key,
-		)
+		kvEvent(ctx, b.Backend, "metadata-delete", key, "metadata/"+key, 2)
 		return nil, err
 	}
 }